Move Yolov11 utils to common place
