+++ models/demos/yolov11/demo/demo.py
@@ -11,8 +11,8 @@ import torch
 from loguru import logger
 
 import ttnn
+from models.demos.utils.common_demo_utils import LoadImages, load_coco_class_names
 from models.demos.yolov11.common import YOLOV11_L1_SMALL_SIZE, load_torch_model
-from models.demos.yolov11.demo.demo_utils import LoadImages, load_coco_class_names
 from models.demos.yolov11.reference import yolov11
 from models.demos.yolov11.runner.performant_runner import YOLOv11PerformantRunner
 from models.demos.yolov11.tt.common import get_mesh_mappers
